MAYBE magit-branch-merged-p: depart from git's behavior
Git has no "is it merged into something appropriate command".
The closest it gets to that is `git branch -d BRANCH', which checks
whether BRANCH is merged either into its upstream or `HEAD'.  We
depart from that and instead check whether BRANCH is merged into
the current branch, because it does not seem safe to assume that
commits that are reachable from the detached `HEAD' won't soon
be dangling if not referenced elsewhere.